Split Workflows.js into separate files
