+++ lib/main.js
@@ -2,20 +2,13 @@
 'use strict';
 
 const packageJson = require('../package.json');
-const Workflows = require('./Workflows');
-const workflowCommands = Object.keys(Workflows);
-const yargs = require('yargs');
 
-const argv = yargs
+require('yargs')
   .usage('Usage: $0 <command> [options]')
-  .demand(1, 'Available commands: ' + Object.keys(Workflows).join(', '))
+  .command(require('./options/init.js'))
+  .command(require('./options/serve.js'))
+  .command(require('./options/static.js'))
+  .command(require('./options/validate.js'))
   .epilogue(`See ${packageJson.homepage} for help.`)
-  .check(function(_argv) {
-    if (_argv._[0] && !!~workflowCommands.indexOf(_argv._[0])) {
-      return true;
-    }
-    throw new Error(`'${_argv._[0]}' is not a valid option! Available commands: ${workflowCommands.join(', ')}`);
-  })
-  .argv;
-
-Workflows[argv._[0]](argv._.slice(1));
+  .demand(1, 'One of the above commands is required.')
+  .strict().argv;
