+++ package.json
@@ -30,14 +30,15 @@
     "css-loader": "^0.23.1",
     "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "0.8.5",
+    "fs-extra": "^0.30.0",
     "invariant": "2.2.1",
     "json-loader": "^0.5.4",
-    "ncp": "2.0.0",
     "postcss-loader": "^0.9.1",
     "raw-loader": "^0.5.1",
     "react": "^15.1.0",
     "react-dom": "^15.1.0",
     "react-hot-loader": "^3.0.0-beta.2",
+    "readline-sync": "^1.4.2",
     "style-loader": "^0.13.1",
     "svgo-loader": "^1.1.0",
     "temp": "0.8.3",
